refactor(client): namespace transport modules
